Feat(Mail): Add button to resend confirmation emails
Buttons have been added to resend the confirmation email either individually or collectively to signees who are still not verified. Also added an endpoint to check how many signees would be affected by the resubmission (in case a limit is present on many emails can be sent by the service adopted by the user).